Refactor checkpoint helpers and remove lint dead code
