+++ src/r2dreamer/launch/parity/train_parity.py
@@ -13,7 +13,6 @@ import os
 import sys
 import time
 
-import numpy as np
 
 ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..", "..", ".."))
 EXT = os.path.join(ROOT, "external", "r2dreamer")
