+++ src/r2dreamer/manifest.py
@@ -13,7 +13,6 @@ import os
 import platform
 import socket
 import subprocess
-import sys
 from datetime import datetime, timezone
 from pathlib import Path
 from typing import Any
