+++ src/vggt/jax/rope.py
@@ -13,7 +13,6 @@ jit-friendly without a stateful frequency cache.
 
 from __future__ import annotations
 
-import jax
 import jax.numpy as jnp
