+++ tests/r2dreamer/test_optim.py
@@ -3,7 +3,6 @@
 import jax
 import jax.numpy as jnp
 import optax
-import pytest
 
 from src.shared.optim import laprop, agc
