+++ tests/vggt/test_jax_integration.py
@@ -8,7 +8,6 @@ PyTorch extractor to within the production tolerance.
 
 from __future__ import annotations
 
-import os
 from pathlib import Path
 
 import jax
